backend/requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.4.1
Babel==2.9.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.6
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
docutils==0.17.1
drf-extra-fields==3.1.1
fonttools==4.26.1
gunicorn==20.1.0
idna==3.2
imagesize==1.2.0
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
recommonmark==0.7.1
reportlab==3.6.0
requests==2.26.0
requests-oauthlib==1.3.0
rinoh-typeface-dejavuserif==0.1.3
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.5.3
six==1.16.0
snowballstemmer==2.1.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
